Updated to run on EC2

+++ Dockerfile
@@ -13,21 +13,22 @@ RUN apt-get update -q && \
         libxext6 \
         libxrender1 \
         wget \
-        zip \
+        cron \
+        awscli \
     && apt-get clean
 
 ENV PATH /opt/conda/bin:$PATH
 
 # Leave these args here to better use the Docker build cache
-ARG CONDA_VERSION=py37_4.8.2
+ARG CONDA_VERSION=py38_4.9.2
 ARG CONDA_MD5=122c8c9beb51e124ab32a0fa6426c656
 
 RUN wget --quiet https://repo.anaconda.com/miniconda/Miniconda3-${CONDA_VERSION}-Linux-x86_64.sh -O miniconda.sh && \
-    # echo "${CONDA_MD5}  miniconda.sh" > miniconda.md5 && \
-    # if ! md5sum --status -c miniconda.md5; then exit 1; fi && \
+    echo "${CONDA_MD5}  miniconda.sh" > miniconda.md5 && \
+    if ! md5sum --status -c miniconda.md5; then exit 1; fi && \
     mkdir -p /opt && \
     sh miniconda.sh -b -p /opt/conda && \
-    # rm miniconda.sh miniconda.md5 && \
+    rm miniconda.sh miniconda.md5 && \
     ln -s /opt/conda/etc/profile.d/conda.sh /etc/profile.d/conda.sh && \
     echo ". /opt/conda/etc/profile.d/conda.sh" >> ~/.bashrc && \
     echo "conda activate base" >> ~/.bashrc && \
@@ -45,13 +46,12 @@ RUN conda install -c conda-forge requests && \
     conda install -c conda-forge pandas && \
     conda install -c conda-forge xarray
 
-COPY index.py   ./
+COPY grib2ToCovjson.py   ./
 COPY converter   ./
+COPY crontab   /etc/cron.d/crontab
 
-RUN cd /opt/conda/lib/python3.7/site-packages && \
-    zip -r ../../../../../my-deployment-package.zip . && \
-    cd ../../../../../ && \
-    zip -g my-deployment-package.zip index.py && \
-    zip -r my-deployment-package.zip pycovjson
+RUN chmod 0644 /etc/cron.d/crontab && \
+    crontab /etc/cron.d/crontab && \
+    chmod 0744 ./grib2ToCovjson.py
 
-# CMD ["python", "index.handler"]
+CMD ["cron", "-f"]
